feat(db): infer DB from schema in createDbClient + KickDbRegister augmentation (M2.A-T6)
Adopters drop the manually-maintained `interface DB` and the
`KickDbClient<MyDb>` cast. Two changes carry the type story:

* `createDbClient<TSchema, DB = SchemaToKysely<TSchema>>` — the returned
  client is typed directly from the schema parameter, breaking the
  `dbClient → RegisteredDB → KickDbRegister['db'] → typeof dbClient`
  cycle that would otherwise resolve to `unknown`.
* `KickDbRegister` interface — module-augmentable registry. Adopters
  declare once:
      declare module '@forinda/kickjs-db' {
        interface KickDbRegister { db: typeof dbClient }
      }
  and `KickDbClient` (no explicit generic) widens everywhere via the
  Register-driven default. Named `KickDbRegister` rather than a generic
  `Register` to avoid collisions with other libraries augmenting a
  same-named interface.

The task-kickdb-api example proves the win end-to-end:
* drops the hand-written `interface Schema` and `KickDbClient<Schema>`
  cast in `db/client.ts`;
* adds `db/register.ts` declaring the augmentation;
* repositories use bare `KickDbClient` and the schema-derived shape
  flows in via DI.

The example schema is also split into a `db/schema/` folder with a
barrel `index.ts` exercising the multi-file path: per-table modules
(`users.ts`, `workspaces.ts`, `tasks.ts`) and a separate `relations.ts`
to avoid an import cycle. SchemaToKysely<typeof schema> filters
non-table exports out of `keyof DB` automatically.
