+++ examples/task-kickdb-api/src/db/schema/index.ts
@@ -1,0 +1,10 @@
+// Barrel — `import * as schema from './schema'` resolves here. Splitting
+// tables across files exercises the multi-file SchemaToKysely<typeof schema>
+// path: per-table modules contribute their own table refs, relations live
+// alongside but are filtered out of `keyof DB` by the SchemaToKysely<S>
+// `extends TableDecl<...>` clause.
+
+export * from './users'
+export * from './workspaces'
+export * from './tasks'
+export * from './relations'
