+++ examples/task-kickdb-api/src/modules/workspaces/workspaces.repository.ts
@@ -1,8 +1,6 @@
 import { Service, Inject } from '@forinda/kickjs'
 import { DB_PRIMARY, type KickDbClient } from '@forinda/kickjs-db'
 
-import type { Db } from '../../db/client'
-
 export interface NewWorkspace {
   name: string
   slug: string
@@ -14,27 +12,23 @@ export interface NewWorkspace {
 export class WorkspacesRepository {
   constructor(@Inject(DB_PRIMARY) private readonly db: KickDbClient) {}
 
-  private get typed(): Db {
-    return this.db as Db
+  list() {
+    return this.db.selectFrom('workspaces').selectAll().orderBy('createdAt', 'asc').execute()
   }
 
-  async list(): Promise<unknown[]> {
-    return this.typed.selectFrom('workspaces').selectAll().orderBy('createdAt', 'asc').execute()
+  findById(id: string) {
+    return this.db.selectFrom('workspaces').selectAll().where('id', '=', id).executeTakeFirst()
   }
 
-  async findById(id: string): Promise<unknown | undefined> {
-    return this.typed.selectFrom('workspaces').selectAll().where('id', '=', id).executeTakeFirst()
-  }
-
-  async create(input: NewWorkspace): Promise<unknown> {
-    return this.typed
+  create(input: NewWorkspace) {
+    return this.db
       .insertInto('workspaces')
       .values({
         name: input.name,
         slug: input.slug,
         description: input.description ?? null,
         ownerId: input.ownerId,
-      } as never)
+      })
       .returningAll()
       .executeTakeFirstOrThrow()
   }
