+++ packages/db/src/client/register.ts
@@ -3,13 +3,13 @@
  * declare:
  *
  *   declare module '@forinda/kickjs-db' {
- *     interface Register {
+ *     interface KickDbRegister {
  *       db: typeof appDbClient
  *     }
  *   }
  *
  * Once the augmentation is in scope, `KickDbClient` (with no explicit
- * generic) widens to `KickDbClient<KickDbSchemaFromRegister>` everywhere —
+ * generic) widens to the registered DB shape everywhere —
  * `@Inject(DB_PRIMARY) private db!: KickDbClient` produces the typed
  * client with no manual cast at the call site.
  *
@@ -17,18 +17,21 @@
  * `.kickjs/types/kick__db.d.ts` for adopters who opt into codegen. Adopters
  * who prefer to declare it by hand do so once in any module that's reached
  * by their tsconfig.
+ *
+ * Named `KickDbRegister` rather than a generic `Register` to avoid
+ * collisions with other libraries that augment a same-named interface.
  */
 // eslint-disable-next-line @typescript-eslint/no-empty-object-type
-export interface Register {}
+export interface KickDbRegister {}
 
 /**
- * Pull the registered DB shape out of the augmentable Register interface.
+ * Pull the registered DB shape out of the augmentable KickDbRegister.
  * `unknown` when the adopter hasn't declared an augmentation — keeps the
  * M1-permissive fallback intact.
  */
-export type RegisteredDB = Register extends { db: { kysely: { __DB__?: never } } }
-  ? unknown // Register['db'] is a phantom shape; this branch is unreachable.
-  : Register extends { db: infer D }
+export type RegisteredDB = KickDbRegister extends { db: { kysely: { __DB__?: never } } }
+  ? unknown // KickDbRegister['db'] is a phantom shape; this branch is unreachable.
+  : KickDbRegister extends { db: infer D }
     ? D extends { kysely: import('kysely').Kysely<infer X> }
       ? X
       : unknown
